LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/textIndent-expected.txt


PASS textIndent responsive to style changes
FAIL textIndent responsive to inherited textIndent changes assert_equals: expected "300px hanging each-line" but got "200px hanging each-line"
